requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-WTF>=1.1.0
Pillow>=10.0.0
python-dotenv==1.0.0
Werkzeug>=3.0.0
email-validator>=2.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.7
Flask-Mail==0.9.1 